GAIA/requirements.txt
Flask==2.0.1
anthropic==0.3.0
pandas==1.5.3
python-dotenv==1.0.0
requests==2.28.2
flask-login==0.6.0
flask-wtf==1.1.0